Update 'Admin' model

+++ codecraft-web/Controllers/AdminsController.cs
@@ -26,7 +26,7 @@ namespace codecraft_web.Controllers
         }
 
         // GET: Admins/Details/5
-        public async Task<IActionResult> Details(long? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -66,7 +66,7 @@ namespace codecraft_web.Controllers
         }
 
         // GET: Admins/Edit/5
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -86,7 +86,7 @@ namespace codecraft_web.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, [Bind("Id,Username,Password,CreatedAt,UpdatedAt")] Admin admin)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Username,Password,CreatedAt,UpdatedAt")] Admin admin)
         {
             if (id != admin.Id)
             {
@@ -117,7 +117,7 @@ namespace codecraft_web.Controllers
         }
 
         // GET: Admins/Delete/5
-        public async Task<IActionResult> Delete(long? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -137,7 +137,7 @@ namespace codecraft_web.Controllers
         // POST: Admins/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var admin = await _context.Admin.FindAsync(id);
             if (admin != null)
@@ -149,7 +149,7 @@ namespace codecraft_web.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool AdminExists(long id)
+        private bool AdminExists(int id)
         {
             return _context.Admin.Any(e => e.Id == id);
         }
